requirements.txt
pyyaml
unidecode
pypdf
pylatex
pandas
bibtexparser
